added col validation to map_continents + electricity map with continent to v3.1
